pyproject.toml
[tool.poetry]
name = "music_update_bot"
version = "0.1.0"
description = "Бот для ведения тг каналов с дискографией арктистов."
authors = ["Greatsol <ilia2309ilia@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^13.14"
requests = "^2.28.1"
yandex-music = "^2.0.1"
black = "^22.8.0"
pylint = "^2.15.3"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
